shyftdb running in parallel with ldb
